requirements.txt
apns2==0.7.2
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.1.7
django-environ==0.4.5
django-push-notifications==2.0.0
factory-boy==3.2.0
Faker==7.0.1
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
py-vapid==1.8.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pywebpush==1.13.0
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
twilio==6.55.0
urllib3==1.26.4
